server/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
tensorflow-cpu==2.13.0
numpy==1.24.3
Pillow==10.0.0
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
motor==3.3.2
typing-extensions==4.5.0
pydantic[email]==1.10.13